Refactor Markdown compilation into its own Eleventy Plugin
